refactor: normalize plugin result
