+++ crates/rolldown/src/bundler/mod.rs
@@ -6,7 +6,5 @@ mod graph;
 mod module;
 mod module_loader;
 pub mod options;
-#[allow(dead_code)]
-pub mod plugin;
 mod resolve_id;
 mod visitors;
