+++ crates/rolldown/src/plugin/plugin_driver.rs
